requirements.txt
flask
psutil
transformers
sentence-transformers
pandas
torch
llama-cpp-python
requests
bme680
groq
